+++ src/utils/file_operations.py
@@ -1,0 +1,6 @@
+import os
+
+
+def count_files(directory: str) -> int:
+    total_files: int = sum([len(files) for _, _, files in os.walk(directory)])
+    return total_files
